+++ src/system/run.js
@@ -135,10 +135,8 @@ var Run = {
           .value();
 
         if (!_.isEmpty(port_data)) {
-          var retry   = options.wait.retry   || config('docker:run:retry');
           var timeout = options.wait.timeout || config('docker:run:timeout');
-
-          yield this._wait_available(system, port_data, container, retry, timeout);
+          yield this._wait_available(system, port_data, container, timeout);
         }
       }
 
@@ -248,7 +246,7 @@ var Run = {
   },
 
   // Wait for container/system available
-  _wait_available(system, port_data, container, retry, timeout) {
+  _wait_available(system, port_data, container, timeout) {
     return async(this, function* () {
       var host;
       if (config('agent:requires_vm')) {
@@ -274,7 +272,7 @@ var Run = {
       }));
 
       var address = `tcp://${host}:${port_data.port}`;
-      var running = yield net.waitService(address, retry, wait_opts);
+      var running = yield net.waitService(address, wait_opts);
 
       if (!running) {
         var data = yield container.inspect();
@@ -283,7 +281,6 @@ var Run = {
         var log = t('errors.run_timeout_error', {
           system: system.name,
           port: port_data && port_data.port,
-          retry: retry,
           timeout: timeout,
           hostname: system.url.underline,
         });