+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## dev
+
+* Enhancements
+  * [Manifest] `wait` accepts a simple number #148
+
 ## v0.14.4 - (2015-06-26)
 
 * Bug